% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3plus-package.R
\docType{package}
\name{d3plus2}
\alias{d3plus2}
\alias{d3plus2-package}
\title{An `htmlwidget` interface to the
d3plus javascript chart library}
\description{
An `htmlwidget` interface to the
d3plus javascript chart library
}
\author{
Mauricio Vargas
}
