% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3plusBubbles.R
\name{d3plusBubblesOutput}
\alias{d3plusBubblesOutput}
\title{Widget output function for use in Shiny}
\usage{
d3plusBubblesOutput(outputId, width = "100\%", height = "500px")
}
\description{
Widget output function for use in Shiny
}

