% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3plusLines.R
\name{d3plusLinesOutput}
\alias{d3plusLinesOutput}
\title{Widget output function for use in Shiny}
\usage{
d3plusLinesOutput(outputId, width = "100\%", height = "500px")
}
\description{
Widget output function for use in Shiny
}

