% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3plusNetwork.R
\name{d3plusNetworkOutput}
\alias{d3plusNetworkOutput}
\title{Widget output function for use in Shiny}
\usage{
d3plusNetworkOutput(outputId, width = "100\%", height = "500px")
}
\description{
Widget output function for use in Shiny
}

