% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3plus.R
\name{d3plusOutput}
\alias{d3plusOutput}
\title{Widget output function for use in Shiny}
\usage{
d3plusOutput(outputId, width = "100\%", height = "500px")
}
\description{
Widget output function for use in Shiny
}

