% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3plusRingsFocus.R
\name{d3plusRingsFocusOutput}
\alias{d3plusRingsFocusOutput}
\title{Widget output function for use in Shiny}
\usage{
d3plusRingsFocusOutput(outputId, width = "100\%", height = "500px")
}
\description{
Widget output function for use in Shiny
}

