% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d3plusRings.R
\name{d3plusRingsOutput}
\alias{d3plusRingsOutput}
\title{Widget output function for use in Shiny}
\usage{
d3plusRingsOutput(outputId, width = "100\%", height = "500px")
}
\description{
Widget output function for use in Shiny
}

