% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d3plusScatter.R
\name{d3plusScatter}
\alias{d3plusScatter}
\title{<Add Title>}
\usage{
d3plusScatter(data = NULL, axis = c(2, 3), options = list(),
  width = NULL, height = NULL)
}
\description{
<Add Description>
}

