% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3plusScatter.R
\name{d3plusScatterOutput}
\alias{d3plusScatterOutput}
\title{Widget output function for use in Shiny}
\usage{
d3plusScatterOutput(outputId, width = "100\%", height = "500px")
}
\description{
Widget output function for use in Shiny
}

