% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3plusTree.R
\name{d3plusTreeOutput}
\alias{d3plusTreeOutput}
\title{Widget output function for use in Shiny}
\usage{
d3plusTreeOutput(outputId, width = "100\%", height = "500px")
}
\description{
Widget output function for use in Shiny
}

