% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3plus.R
\name{renderD3plus}
\alias{renderD3plus}
\title{Widget render function for use in Shiny}
\usage{
renderD3plus(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

