% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3plusBubbles.R
\name{renderD3plusBubbles}
\alias{renderD3plusBubbles}
\title{Widget render function for use in Shiny}
\usage{
renderD3plusBubbles(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

