% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3plusLines.R
\name{renderD3plusLines}
\alias{renderD3plusLines}
\title{Widget render function for use in Shiny}
\usage{
renderD3plusLines(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

