% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3plusNetwork.R
\name{renderD3plusNetwork}
\alias{renderD3plusNetwork}
\title{Widget render function for use in Shiny}
\usage{
renderD3plusNetwork(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

