% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3plusRings.R
\name{renderD3plusRings}
\alias{renderD3plusRings}
\title{Widget render function for use in Shiny}
\usage{
renderD3plusRings(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

