% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3plusRingsFocus.R
\name{renderD3plusRingsFocus}
\alias{renderD3plusRingsFocus}
\title{Widget render function for use in Shiny}
\usage{
renderD3plusRingsFocus(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

