% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d3plusScatter.R
\name{renderD3plusScatter}
\alias{renderD3plusScatter}
\title{Widget render function for use in Shiny}
\usage{
renderD3plusScatter(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

