% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3plusTree.R
\name{renderD3plusTree}
\alias{renderD3plusTree}
\title{Widget render function for use in Shiny}
\usage{
renderD3plusTree(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

